requirements.txt
numpy
torch
torchvision
Pillow
matplotlib
nltk
rouge-score
joeynmt
ipython
importlib-metadata
